backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-cors==4.0.0
flask-jwt-extended==4.5.2
pymetta==0.1.1
python-dotenv==1.0.0
werkzeug==2.3.4
marshmallow==3.19.0
gunicorn==21.2.0
pytest==7.3.1